requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
apturl==0.5.2
asgiref==3.8.1
astd==6.5
async-timeout==4.0.3
attrs==23.2.0
bandit==1.6.2
bcrypt==3.2.0
billiard==4.2.0
blinker==1.4
Brlapi==0.8.3
celery==5.3.6
certifi==2020.4.5.1
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.4
command-not-found==0.3
crispy-bootstrap5==2024.2
cryptography==2.9.2
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.6.0
diff-match-patch==20181111
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.6
django-active-link==0.1.6
django-allauth==0.61.1
django-appconf==1.0.4
django-compressor==2.4
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-crontab==0.7.1
django-datatables-view==1.18.0
django-developer-panel==0.1.2
django-discover-runner==1.0
django-excel==0.0.10
django-extensions==1.7.9
django-import-export==1.2.0
django-pandas==0.6.1
django-role-permissions==2.2.1
django-session-security==2.6.6
django-session-timeout==0.1.0
django-simple-history==2.10.0
django-template-repl==0.3.0
django-timezone-field==3.0
django-tz-detect==0.2.8
django-webpush==0.3.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
duplicity==0.8.21
et-xmlfile==1.0.1
fasteners==0.14.1
firebase==4.0.1
frozenlist==1.4.1
future==0.18.2
getmac==0.9.4
gitdb==4.0.2
GitPython==3.1.0
http-ece==1.1.0
httplib2==0.20.2
idna==2.9
importlib-metadata==4.6.4
jdcal==1.4.1
jeepney==0.7.1
keyring==23.5.0
kombu==5.3.6
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lml==0.0.9
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
mailtrap==2.0.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.5
netifaces==0.11.0
numpy==1.26.3
oauthlib==3.2.0
odfpy==1.4.1
olefile==0.46
openpyxl==3.1.2
pandas==1.3.5
paramiko==2.9.3
paywix==2.0.2
pbr==5.4.4
pexpect==4.8.0
Pillow==7.1.1
prompt-toolkit==3.0.43
protobuf==3.12.4
psycopg==3.1.17
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
py-vapid==1.7.1
pycairo==1.20.1
pycparser==2.20
pycryptodomex==3.11.0
pycups==2.0.1
pyexcel==0.5.14
pyexcel-io==0.5.18
pyexcel-ods==0.5.6
pyexcel-webio==0.1.4
pyexcel-xls==0.5.8
pyexcel-xlsx==0.5.7
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
PyOpenGL==3.1.5
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-firebase==1.2
python-nmap==0.7.1
python3-openid==3.2.0
pytz==2019.3
pywebpush==1.9.4
pyxattr==0.7.2
pyxdg==0.27
pyxlsb==1.0.6
PyYAML==5.3.1
rcssmin==1.0.6
redis==5.0.3
reportlab==3.6.8
requests==2.23.0
requests-oauthlib==1.4.0
rjsmin==1.1.0
schedule==1.2.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
simplejson==3.17.6
six==1.14.0
smmap==3.0.1
sqlparse==0.3.1
ssh-import-id==5.11
stevedore==1.32.0
systemd-python==234
tablib==1.1.0
texttable==1.6.2
twilio==9.0.0
typing_extensions==4.9.0
tzdata==2023.4
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
vidcutter==6.0.5.1
vine==5.1.0
wadllib==1.3.6
wcwidth==0.2.13
xdg==5
xkit==0.0.0
xlrd==1.2.0
XlsxWriter==1.2.8
xlwt==1.3.0
yapf==0.29.0
yarl==1.9.4
youtube-dl==2021.12.17
zipp==1.0.0
